also fix tl_form_field.accesskey
